# internal/api/v1/spec/openapi.yaml
openapi: 3.1.0
info:
  title: gosidian REST API
  version: "v1"
  description: |
    REST/JSON API consumed by the gosidian Vue 3 SPA (v2.0). Audience is
    the browser; the parallel MCP transport (`/mcp/sse`) covers agents.

    Auth: every endpoint outside the `public` tag requires a Bearer
    token in the `Authorization` header. Tokens are issued by
    `POST /api/v1/login` and managed by an SPA-specific token store
    distinct from the long-lived MCP token surface.

    Phase status: this spec is grown phase by phase together with
    handler implementations. Endpoints whose body is not yet wired
    return `501 Not Implemented` with `error.code = server.unavailable`
    so the SPA can render a "feature coming soon" state without
    special-casing each fetch.
  contact:
    name: gosidian
    url: https://github.com/daniele-chiappa/gosidian
  license:
    name: MIT
servers:
  - url: /
    description: same-origin
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: gsp_<base64url>
      description: |
        Bearer token issued by `POST /api/v1/login`. Sliding TTL 24h,
        absolute hard expiry 7d. Refresh via `POST /api/v1/refresh`
        before `expires_at`; full re-login required after the hard
        expiry regardless of refresh activity.
...
security:
  - bearer: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/folders/{path}/export.zip:
    get:
      tags: [tree]
      summary: Zip of a folder (the tree's context menu).
      description: >-
        Every file under the folder, as the project export: hidden files,
        node_modules and symlinks left out, names starting at the folder
        itself (docs/... for Work/a/docs). Any signed-in account that can
        read the project; 10 exports per 10 minutes per account, shared with
        the project and vault exports; audited as export.
      parameters:
        - in: path
          name: path
          required: true
          description: Vault-relative folder path; its slashes stay as they are.
          schema:
            type: string
      responses:
        "200":
          description: The zip, streamed.
          content:
            application/zip: {}
        "403":
          description: The anonymous guest of open mode.
        "404":
          description: Not a folder, or a project the account cannot read.
        "429":
          description: Over the export limit (Retry-After).
  /api/v1/folders/{path}:
    delete:
      tags: [tree]
      summary: Move a folder inside a project to the trash.
      description: >-
        The folder and all it holds become one trash entry, restored in one
        go from /api/v1/trash like a note: write access on the project. A
        project folder is deleted with DELETE /api/v1/projects/{name}.
        Audited as delete.
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trashed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  trash_id: { type: string }
                  removed:
                    type: array
                    items: { type: string }
                    description: The notes the folder held, now out of the index.
        "400":
          description: A project folder, a hidden path, or a folder that is, holds or sits in the server's state directory.
        "403":
          description: No write access on the project.
        "404":
          description: Not a folder, or a project the account cannot read.
        "503":
          description: The trash is not enabled; a folder is never deleted for good this way.
...